+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/CredentialRequestProof.kt
@@ -22,6 +22,8 @@ data class CredentialRequestProof(
     /**
      * OID4VCI: A CWT (RFC8392) is used as proof of possession. When [proofType] is `cwt`, a proof object MUST include
      * a `cwt` claim containing a CWT defined in Section 7.2.1.3.
+     *
+     * Removed in OID4VCI Draft 14, kept here for a bit of backwards-compatibility
      */
     @SerialName("cwt")
     val cwt: String? = null,