+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Release NEXT:
  * Add `KeyStoreMaterial` to JVM target for convenience
  - Update implementation of [OpenID for Verifiable Credential Issuance](https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html) to draft 14 from 2024-08-21
  - Move some fields from `IssuerMetadata` to `OAuth2AuthorizationServerMetadata` to match the semantics
+ - Remove proof type `cwt` for OpenID for Verifiable Credential Issuance, as per draft 14
 
 Release 4.1.2:
  * In `OidcSiopVerifier` add parameter `nonceService` to externalize creation and validation of nonces, e.g. for deployments in load-balanced environments